chore: clean up and tests
