+++ pkg/rewards/types.go
@@ -1,6 +1,11 @@
 package rewards
 
-import "math/big"
+import (
+	"math/big"
+
+	"github.com/Layr-Labs/eigenlayer-cli/pkg/types"
+	gethcommon "github.com/ethereum/go-ethereum/common"
+)
 
 type RewardResponse struct {
 	Rewards []Reward `json:"rewards"`
@@ -36,4 +41,45 @@ type UnclaimedRewardResponse struct {
 type NormalizedUnclaimedReward struct {
 	TokenAddress string   `csv:"tokenAddress"`
 	WeiAmount    *big.Int `csv:"weiAmount"`
+}
+
+type ClaimConfig struct {
+	Network                   string
+	RPCUrl                    string
+	EarnerAddress             gethcommon.Address
+	RecipientAddress          gethcommon.Address
+	ClaimerAddress            gethcommon.Address
+	Output                    string
+	OutputType                string
+	Broadcast                 bool
+	TokenAddresses            []gethcommon.Address
+	RewardsCoordinatorAddress gethcommon.Address
+	ClaimTimestamp            string
+	ChainID                   *big.Int
+	ProofStoreBaseURL         string
+	Environment               string
+	SignerConfig              *types.SignerConfig
+}
+
+type SetClaimerConfig struct {
+	ClaimerAddress            gethcommon.Address
+	Network                   string
+	RPCUrl                    string
+	Broadcast                 bool
+	RewardsCoordinatorAddress gethcommon.Address
+	ChainID                   *big.Int
+	SignerConfig              *types.SignerConfig
+	EarnerAddress             gethcommon.Address
+	Output                    string
+	OutputType                string
+}
+
+type ShowConfig struct {
+	EarnerAddress gethcommon.Address
+	NumberOfDays  int64
+	Network       string
+	Environment   string
+	ClaimType     ClaimType
+	ChainID       *big.Int
+	Output        string
 }
